+++ examples/resources/iosxe_line/resource.tf
@@ -1,20 +1,20 @@
 resource "iosxe_line" "example" {
   console = [
     {
-      first                  = "0"
-      exec_timeout_minutes   = 45
-      exec_timeout_seconds   = 25
-      privilege_level_number = 15
-      stopbits               = "1"
-      password_type          = "0"
-      password_secret        = "testpasswd"
+      first                = "0"
+      exec_timeout_minutes = 45
+      exec_timeout_seconds = 25
+      privilege_level      = 15
+      stopbits             = "1"
+      password_type        = "0"
+      password             = "testpasswd"
     }
   ]
   vty = [
     {
       first = 10
       last  = 27
-      access_class = [
+      access_classes = [
         {
           direction   = "in"
           access_list = "2"
@@ -24,9 +24,9 @@ resource "iosxe_line" "example" {
       exec_timeout_minutes         = 45
       exec_timeout_seconds         = 25
       password_type                = "0"
-      password_secret              = "testpasswd"
+      password                     = "testpasswd"
       transport_preferred_protocol = "none"
-      escape_character_char        = "27"
+      escape_character             = "27"
     }
   ]
 }